--- app/blog article components.js
@@ -1,34 +1,0 @@
-<div className='h-auto'>
-    <div className='container justify-center h-auto m-auto w-auto max-w-md sm:max-w-xl border-solid border-black outline-4 border-2 rounded-lg'>
-        <div className='flex flex-row text-black text-xs m-5'>
-            Tag 1
-            Tag 2
-        </div>
-        <div className='m-5'>
-            <Image
-                src='/monkey2.webp'
-                width={600}
-                height={400}
-                alt="Picture of the author"
-                className='rounded-xl'
-            />
-        </div>
-        <h1 className='text-xl text-black m-5 truncate'>
-            Amet enim esse ex Lorem commodo commodo voluptate est ipsum ipsum pariatur enim aliqua enim.
-        </h1>
-        <div className='flex flex-col textarea-bordered m-5'>
-            <div className='container h-52 overflow-hidden text-ellipsis mb-5'>
-                <h5 className='textarea-bordered text-black'>
-                    Content9
-                    Ea pariatur labore in ex consectetur do in et veniam. Ea voluptate aliquip nostrud est cupidatat veniam nisi fugiat est sint est amet. Dolor incididunt mollit velit velit adipisicing laborum commodo.
-                    Incididunt ea ad excepteur non excepteur in ipsum eiusmod pariatur. Magna laborum cillum sit ad laborum nisi elit voluptate in exercitation est incididunt qui. Enim dolore dolor nostrud Lorem amet ea sint veniam incididunt proident. Sunt dolore et quis commodo mollit esse commodo labore eiusmod sunt enim deserunt. Ullamco minim sit proident tempor sit esse adipisicing. Est occaecat dolore in commodo ullamco voluptate cillum commodo ad elit est esse.
-                    Nostrud nostrud reprehenderit sint consequat excepteur ex. Fugiat culpa velit ipsum aliqua nulla mollit et adipisicing Lorem ullamco aliqua quis velit. Lorem eu proident mollit quis officia enim quis labore proident. In duis pariatur aliqua cupidatat officia esse. Ex fugiat incididunt exercitation ex aliqua laborum. Ex consequat proident deserunt ullamco minim Lorem qui. Excepteur velit occaecat ipsum culpa pariatur proident proident enim deserunt magna ad.
-                    Ex exercitation sint amet fugiat ipsum quis. Velit labore aliquip veniam eiusmod aliquip. Voluptate occaecat incididunt ullamco reprehenderit cillum. Amet sunt irure irure aliqua minim deserunt nostrud ea qui et voluptate reprehenderit enim culpa.
-                </h5>
-            </div>
-        </div>
-        <div className='container flex justify-end w-auto m-2'>
-            <button className='btn btn-xs sm:btn-sm md:btn-md lg:btn-lg w-32 relative'>Read More</button>
-        </div>
-    </div>
-</div>
