+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "npx tailwindcss -i ./input.css -o ./dist/output.css --watch",
+    "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
